requirements.txt
# Face Recognition System - Exact Version Requirements
# Tested with Python 3.12.7

# Core face detection and recognition
ultralytics==8.3.245
huggingface-hub==0.36.0
supervision==0.27.0
face-recognition==1.3.0
retina-face==0.0.17
deepface==0.0.96

# Image processing
numpy==1.26.4
Pillow==10.3.0
opencv-python==4.11.0.86

# Deep learning frameworks
torch==2.9.1
torchvision==0.24.1

# Speech recognition and audio (Distil-Whisper) - Optional
# Install these for speech-to-text features
transformers>=4.39.0
accelerate>=0.20.0
datasets[audio]>=2.14.0
soundfile>=0.12.0
pyaudio>=0.2.14

# Gemini API - Optional
# Install for AI voice chat features
google-generativeai>=0.3.0
